feat: centralize system observability diagnostics (PR 11)
